# swagger/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Pattern Go RestAPI + Socket.io v2. with the same ports.",
        "title": "RestAPI \u0026 Socket.io v2",
        "termsOfService": "https://id.linkedin.com/public-profile/in/arditya-kekasi",
        "contact": {
            "name": "Arditya Kekasi",
            "url": "http://www.kekasi.co.id",
            "email": "arditya@kekasi.co.id"
        },
        "license": {
            "name": "YouTube KekasiGen",
            "url": "https://kekasi.link/kekasigensub"
        },
        "version": "1.0"
    },
    "paths": {
        "/3rd/product/all": {
            "get": {
                "description": "https://dummyjson.com/docs/products#all",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "3rd Party"
                ],
                "responses": {}
            }
        },
        "/3rd/product/{id}": {
            "get": {
                "description": "https://dummyjson.com/docs/products#single",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "3rd Party"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Product ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/keluarga/add": {
            "post": {
                "description": "2.a Dapat menambahkan data orang baru ke keluarga (baru)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Keluarga"
                ],
                "parameters": [
                    {
                        "description": "Tambah Keluarga",
                        "name": "Keluarga",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.ReqAddKeluarga"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    }
                }
            }
        },
        "/keluarga/asset/add": {
            "post": {
                "description": "2.d Dapat menambah data aset keluarga \u0026\u0026 2.g Dapat menambahkan aset yang dimiliki oleh suatu orang dalam keluarga",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Keluarga"
                ],
                "parameters": [
                    {
                        "description": "Tambah Asset Keluarga",
                        "name": "Keluarga",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.ReqAddAssetKeluarga"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    }
                }
            }
        },
        "/keluarga/asset/delete": {
            "post": {
                "description": "2.f Dapat menghapus data aset keluarga \u0026\u0026 2.h Dapat mengurangi asset yang dimiliki oleh suatu orang dalam keluarga",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Keluarga"
                ],
                "parameters": [
                    {
                        "description": "hapus Asset Keluarga",
                        "name": "Keluarga",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.ReqDeletessetKeluarga"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    }
                }
            }
        },
        "/keluarga/asset/riches": {
            "post": {
                "description": "Dapat menampilkan total nilai (price) aset yang dimiliki oleh masing-masing orang dalam suatu keluarga berdasarkan data harga product di api produk berikut https://dummyjson.com/docs/products",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Keluarga"
                ],
                "parameters": [
                    {
                        "description": "Total Asset Keluarga",
                        "name": "Keluarga",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.ReqGetKeluargaAssets"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    }
                }
            }
        },
        "/keluarga/asset/update": {
            "post": {
                "description": "2.e Dapat mengedit data aset keluarga",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Keluarga"
                ],
                "parameters": [
                    {
                        "description": "Ubah Asset Keluarga",
                        "name": "Keluarga",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.ReqUpdatessetKeluarga"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    }
                }
            }
        },
        "/keluarga/delete": {
            "post": {
                "description": "3.c Dapat menghapus data orang dalam keluarga",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Keluarga"
                ],
                "parameters": [
                    {
                        "description": "Hapus Keluarga",
                        "name": "Keluarga",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.ReqDeleteKeluarga"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    }
                }
            }
        },
        "/keluarga/get": {
            "post": {
                "description": "Menampilkan anggota keluarga 1 tingkat di bawah",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Keluarga"
                ],
                "parameters": [
                    {
                        "description": "Pindah Keluarga",
                        "name": "Keluarga",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.ReqGetKeluarga"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    }
                }
            }
        },
        "/keluarga/switch": {
            "post": {
                "description": "2.a Dapat menambahkan data orang baru ke keluarga (Pindah KK)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Keluarga"
                ],
                "parameters": [
                    {
                        "description": "Pindah Keluarga",
                        "name": "Keluarga",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.ReqSwitchKeluarga"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    }
                }
            }
        },
        "/keluarga/update": {
            "post": {
                "description": "2.b Dapat mengedit data orang dalam keluarga",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Keluarga"
                ],
                "parameters": [
                    {
                        "description": "Ubah Keluarga",
                        "name": "Keluarga",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.ReqUpdateKeluarga"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "domain.ReqAddAssetKeluarga": {
            "type": "object",
            "properties": {
                "id_keluarga": {
                    "type": "integer"
                },
                "id_product": {
                    "type": "integer"
                },
                "orang_tua": {
                    "type": "integer"
                }
            }
        },
        "domain.ReqAddKeluarga": {
            "type": "object",
            "required": [
                "nama"
            ],
            "properties": {
                "jenis_kelamin": {
                    "type": "integer"
                },
                "nama": {
                    "type": "string"
                },
                "orang_tua": {
                    "type": "integer"
                }
            }
        },
        "domain.ReqDeleteKeluarga": {
            "type": "object",
            "properties": {
                "id_keluarga": {
                    "type": "integer"
                },
                "orang_tua": {
                    "type": "integer"
                }
            }
        },
        "domain.ReqDeletessetKeluarga": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "id_keluarga": {
                    "type": "integer"
                },
                "id_product": {
                    "type": "integer"
                },
                "orang_tua": {
                    "type": "integer"
                }
            }
        },
        "domain.ReqGetKeluarga": {
            "type": "object",
            "properties": {
                "id_keluarga": {
                    "type": "integer"
                }
            }
        },
        "domain.ReqGetKeluargaAssets": {
            "type": "object",
            "properties": {
                "id_keluarga": {
                    "type": "integer"
                },
                "orang_tua": {
                    "type": "integer"
                }
            }
        },
        "domain.ReqSwitchKeluarga": {
            "type": "object",
            "properties": {
                "id_keluarga": {
                    "type": "integer"
                },
                "orang_tua": {
                    "type": "integer"
                },
                "orang_tua_baru": {
                    "type": "integer"
                }
            }
        },
        "domain.ReqUpdateKeluarga": {
            "type": "object",
            "required": [
                "nama"
            ],
            "properties": {
                "id_keluarga": {
                    "type": "integer"
                },
                "jenis_kelamin": {
                    "type": "integer"
                },
                "nama": {
                    "type": "string"
                },
                "orang_tua": {
                    "type": "integer"
                }
            }
        },
        "domain.ReqUpdatessetKeluarga": {
            "type": "object",
            "properties": {
                "id_keluarga": {
                    "type": "integer"
                },
                "id_product": {
                    "type": "integer"
                },
                "orang_tua": {
                    "type": "integer"
                },
                "product_baru": {
                    "type": "integer"
                }
            }
        },
        "model.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        }
    }
}